+++ apps/vs-code-designer/src/app/commands/deploy/deploy.ts
@@ -278,7 +278,7 @@ async function getLogicAppsPicks(
  * @param context {@link IActionContext}
  * @param fsPath publish path for logic app extension
  */
-async function cleanupPublishBinPath(context: IActionContext, fsPath: string): Promise<void> {
+async function cleanupPublishBinPath(_context: IActionContext, fsPath: string): Promise<void> {
   const netFxWorkerBinPath = path.join(fsPath, 'bin', 'NetFxWorker');
   const netFxWorkerAssetPath = path.join(fsPath, 'NetFxWorker');
   if (await fse.pathExists(netFxWorkerBinPath)) {
@@ -364,7 +364,7 @@ async function getProjectPathToDeploy(
       secret: `@appsetting('${workflowAppAADClientSecret}')`,
     };
 
-    if (parameterizeConnectionsSetting === null || parameterizeConnectionsSetting) {
+    if (parameterizeConnectionsSetting) {
       identityWizardContext?.useAdvancedIdentity
         ? updateAuthenticationParameters(advancedIdentityAuthValue)
         : updateAuthenticationParameters(authValue);