+++ apps/vs-code-designer/src/app/commands/workflows/openMonitoringView/openMonitoringViewForLocal.ts
@@ -135,6 +135,11 @@ export default class OpenMonitoringViewForLocal extends OpenMonitoringViewBase {
         await this.resubmitRun();
         break;
       }
+      case ExtensionCommand.logTelemetry: {
+        const eventName = message.data.name ?? message.data.area;
+        ext.telemetryReporter.sendTelemetryEvent(eventName, { ...message.data });
+        break;
+      }
       default:
         break;
     }