+++ libs/data-mapper-v2/src/components/functionConfigurationMenu/inputTab/inputTab.tsx
@@ -1,5 +1,5 @@
-import { Badge, Button, Caption1, Caption2 } from '@fluentui/react-components';
-import { LinkDismissRegular, AddRegular } from '@fluentui/react-icons';
+import { Badge, Button, Caption1, Caption2, Text } from '@fluentui/react-components';
+import { AddRegular, DeleteRegular } from '@fluentui/react-icons';
 import { useDispatch, useSelector } from 'react-redux';
 import { UnboundedInput } from '../../../constants/FunctionConstants';
 import {
@@ -15,7 +15,6 @@ import { getInputName, getInputValue } from '../../../utils/Function.Utils';
 import type { InputOptionProps } from '../inputDropdown/InputDropdown';
 import { InputDropdown } from '../inputDropdown/InputDropdown';
 import { useStyles } from './styles';
-import { mergeStyles } from '@fluentui/react';
 import { isSchemaNodeExtended } from '../../../utils';
 import {
   connectionDoesExist,
@@ -28,11 +27,23 @@ import DraggableList from 'react-draggable-list';
 import InputListWrapper, { type TemplateItemProps, type CommonProps } from './InputList';
 import { useCallback, useMemo } from 'react';
 import { useIntl } from 'react-intl';
+import { InputCustomInfoLabel } from './inputCustomInfoLabel';
 
 export const InputTabContents = (props: {
   func: FunctionData;
   functionKey: string;
 }) => {
+  const intl = useIntl();
+  const resources = useMemo(
+    () => ({
+      ACCEPTED_TYPES: intl.formatMessage({
+        defaultMessage: 'Accepted types: ',
+        id: 'ZgyD93',
+        description: 'Accepted types',
+      }),
+    }),
+    [intl]
+  );
   const connectionDictionary = useSelector((state: RootState) => state.dataMap.present.curDataMapOperation.dataMapConnections);
   const sourceSchemaDictionary = useSelector((state: RootState) => state.dataMap.present.curDataMapOperation.flattenedSourceSchema);
   const functionNodeDictionary = useSelector((state: RootState) => state.dataMap.present.curDataMapOperation.functionNodes);
@@ -94,7 +105,11 @@ export const InputTabContents = (props: {
               <Caption1 className={styles.inputName}>{input.name}</Caption1>
               <Caption2>{input.tooltip ?? input.placeHolder ?? ''}</Caption2>
             </div>
-            <Caption2 className={styles.allowedTypes}>Allowed types: {input.allowedTypes}</Caption2>
+            <Caption2 className={styles.allowedTypesComponent}>
+              <Text className={styles.typesParent}>{resources.ACCEPTED_TYPES}</Text>
+              {input.allowedTypes}
+            </Caption2>
+            <InputCustomInfoLabel />
           </div>
           <div>
             <span className={styles.inputDropdownWrapper}>
@@ -118,7 +133,7 @@ export const InputTabContents = (props: {
             <Button
               className={styles.listButton}
               appearance="transparent"
-              icon={<LinkDismissRegular />}
+              icon={<DeleteRegular />}
               onClick={() => removeConnection(index)}
             />
           </div>
@@ -160,6 +175,11 @@ const UnlimitedInputs = (props: {
         id: '6eDY1H',
         description: 'Optional Keyword',
       }),
+      ADD_INPUT: intl.formatMessage({
+        defaultMessage: 'Add Input',
+        id: 'wx/ZQP',
+        description: 'Add Input',
+      }),
     }),
     [intl]
   );
@@ -185,13 +205,20 @@ const UnlimitedInputs = (props: {
       <div>
         <span className={styles.unlimitedInputHeaderCell} key="input-name">
           <Caption1>{`${inputsFromManifest[0].name}${inputsFromManifest[0].isOptional ? ` (${stringResources.OPTIONAL})` : ''}`}</Caption1>
+          <InputCustomInfoLabel />
         </span>
-        <span className={mergeStyles(styles.unlimitedInputHeaderCell, styles.allowedTypes)} key="input-types">
-          <Caption2>{`${stringResources.ACCEPT_TYPES}${inputsFromManifest[0].allowedTypes}`}</Caption2>
+        <span className={styles.unlimitedInputHeaderCell} key="input-types">
+          <Caption2>
+            <Text className={styles.typesParent}>{stringResources.ACCEPT_TYPES}</Text>
+            {inputsFromManifest[0].allowedTypes}
+          </Caption2>
         </span>
       </div>
       <DraggableList<TemplateItemProps, CommonProps, any>
-        list={Object.entries(functionConnection.inputs).map((input, index) => ({ input: input[1], index }))}
+        list={Object.entries(functionConnection.inputs).map((input, index) => ({
+          input: input[1],
+          index,
+        }))}
         commonProps={{
           functionKey: props.functionKey,
           data: props.func,
@@ -210,7 +237,7 @@ const UnlimitedInputs = (props: {
         className={styles.addButton}
         appearance="transparent"
       >
-        <Caption1>Add Input</Caption1>
+        <Caption1>{stringResources.ADD_INPUT}</Caption1>
       </Button>
     </div>
   );