+++ libs/data-mapper-v2/src/ui/hooks/useAutoLayout.ts
@@ -1,7 +1,7 @@
 import Elk, { type ElkNode } from 'elkjs/lib/elk.bundled.js';
 import { useEffect, useState } from 'react';
 import { type Node, type Edge, type XYPosition, useStore, useReactFlow } from '@xyflow/react';
-import { isFunctionNode } from '../../utils/ReactFlow.Util';
+import { isFunctionNode, panelWidth } from '../../utils/ReactFlow.Util';
 import { useDispatch } from 'react-redux';
 import type { AppDispatch } from '../../core/state/Store';
 import { updateFunctionNodesPosition } from '../../core/state/DataMapSlice';
@@ -18,7 +18,7 @@ export type LayoutAlgorithm = (nodes: Node[], edges: Edge[], options: LayoutOpti
 
 const elk = new Elk();
 
-const elkLayout: LayoutAlgorithm = async (nodes, edges, options) => {
+const elkLayout: LayoutAlgorithm = async (nodes, _edges, options) => {
   const graph = {
     id: 'elk-root',
     layoutOptions: {
@@ -158,7 +158,12 @@ const useAutoLayout = () => {
           nextNodes
             .filter((node) => isFunctionNode(node.id))
             .reduce((acc: Record<string, XYPosition>, node) => {
-              acc[node.id] = node.position;
+              const { x: currentX, y: currentY } = node.position;
+              acc[node.id] = {
+                // Todo: This is a temporary fix for the layout issue. We need to find a better solution.
+                x: currentX < panelWidth ? currentX + panelWidth : currentX,
+                y: currentY,
+              };
               return acc;
             }, {})
         )