+++ libs/designer-ui/src/lib/editor/base/index.tsx
@@ -61,6 +61,7 @@ export interface BaseEditorProps {
   basePlugins?: BasePlugins;
   initialValue: ValueSegment[];
   children?: React.ReactNode;
+  ariaLabel?: string;
   labelId?: string;
   label?: string;
   valueType?: string;
@@ -96,6 +97,7 @@ export const BaseEditor = ({
   basePlugins = {},
   children,
   labelId,
+  ariaLabel,
   tokenPickerButtonProps,
   valueType,
   dataAutomationId,
@@ -188,6 +190,7 @@ export const BaseEditor = ({
             <ContentEditable
               spellCheck={false}
               className={css('editor-input', readonly && 'readonly')}
+              ariaLabel={ariaLabel}
               ariaLabelledBy={labelId}
               ariaDescribedBy={id}
               tabIndex={0}