+++ libs/designer-ui/src/lib/settings/settingsection/settingTokenField.tsx
@@ -39,10 +39,12 @@ import { Label } from '../../label';
 import { EditorLanguage, equals, getPropertyValue, replaceWhiteSpaceWithUnderscore } from '@microsoft/logic-apps-shared';
 import { MixedInputEditor } from '../../mixedinputeditor/mixedinputeditor';
 import { useMemo } from 'react';
+import { useIntl } from 'react-intl';
 
 export interface SettingTokenFieldProps extends SettingProps {
   id?: string;
   value: ValueSegment[];
+  isDynamic?: boolean;
   isLoading?: boolean;
   errorDetails?: { message: string };
   editor?: string;
@@ -100,6 +102,7 @@ export const TokenField = ({
   placeholder,
   readOnly,
   value,
+  isDynamic,
   isLoading,
   errorDetails,
   showTokens,
@@ -116,18 +119,36 @@ export const TokenField = ({
   onFileNameChange,
   getTokenPicker,
   suppressCastingForSerialize,
+  required,
 }: TokenFieldProps) => {
+  const intl = useIntl();
   const dropdownOptions = useMemo(() => getDropdownOptionsFromOptions(editorOptions), [editorOptions]);
   const labelForAutomationId = useMemo(() => replaceWhiteSpaceWithUnderscore(label), [label]);
+
+  const arrayItemLabel = intl.formatMessage(
+    {
+      defaultMessage: '{label} Item',
+      id: 'fBUCrA',
+      description: 'Label for array item',
+    },
+    { label }
+  );
+
+  const defaultArrayItemLabel = intl.formatMessage({
+    defaultMessage: 'Array Item',
+    id: 'gS4Teq',
+    description: 'Label for array item',
+  });
 
   switch (editor?.toLowerCase()) {
     case constants.PARAMETER.EDITOR.ARRAY:
       return (
         <ArrayEditor
+          isRequired={required}
           labelId={labelId}
           arrayType={editorViewModel.arrayType}
           initialMode={editorOptions?.initialMode}
-          labelProps={{ text: label ? `${label} Item` : 'Array Item' }}
+          labelProps={{ text: label ? arrayItemLabel : defaultArrayItemLabel }}
           placeholder={placeholder}
           readonly={readOnly}
           initialValue={editorViewModel.uncastedValue}
@@ -138,7 +159,8 @@ export const TokenField = ({
           onChange={onValueChange}
           dataAutomationId={`msla-setting-token-editor-arrayeditor-${labelForAutomationId}`}
           // Props for dynamic options
-          options={dropdownOptions.length > 0 ? dropdownOptions : undefined}
+          isDynamic={isDynamic}
+          options={dropdownOptions}
           isLoading={isLoading}
           errorDetails={errorDetails}
           onMenuOpen={onComboboxMenuOpen}
@@ -165,6 +187,7 @@ export const TokenField = ({
           tokenPickerButtonProps={tokenpickerButtonProps}
         />
       );
+
     case constants.PARAMETER.EDITOR.CODE: {
       const { language = EditorLanguage.javascript } = editorOptions || {};
       const customCodeEditor = isCustomCode(editor, language);
@@ -191,6 +214,7 @@ export const TokenField = ({
         />
       );
     }
+
     case constants.PARAMETER.EDITOR.COMBOBOX:
       return (
         <Combobox
@@ -249,6 +273,7 @@ export const TokenField = ({
       return (
         <DictionaryEditor
           labelId={labelId}
+          label={label}
           placeholder={placeholder}
           readonly={readOnly}
           initialValue={value}