+++ libs/designer/src/lib/ui/panel/nodeDetailsPanel/tabs/parametersTab/index.tsx
@@ -448,6 +448,7 @@ const ParameterSection = ({
           editor,
           editorOptions,
           tokenEditor: true,
+          isDynamic: dynamicData !== undefined,
           isLoading: dynamicData?.status === DynamicLoadStatus.LOADING,
           errorDetails: dynamicData?.error ? { message: dynamicData.error.message } : undefined,
           validationErrors,